SYNC-1569 Add 'pending wipe' and change status to wiped when client attempts to sync

+++ modules/apps/sync/sync-service/src/main/java/com/liferay/sync/service/impl/SyncDeviceServiceImpl.java
@@ -31,7 +31,7 @@ public class SyncDeviceServiceImpl extends SyncDeviceServiceBaseImpl {
 
 	@Override
 	public SyncDevice registerSyncDevice(
-			String type, int buildNumber, int featureSet, String uuid)
+			String type, long buildNumber, int featureSet, String uuid)
 		throws PortalException {
 
 		User user = getUser();
